move _next_power_of_2 to utils
